🔄 Refactor: Split auth/user routes and controllers + improve auth middleware
