+++ server/middleware/auth.js
@@ -11,7 +11,7 @@ module.exports = function (req, res, next) {
 
     try {
         const decoded = jwt.verify(token, process.env.JWT_SECRET);
-        req.user = decoded; // user_id, role
+        req.user = decoded;
         next();
     } catch (err) {
         return res.status(401).json({ error: 'Token invalide ou expiré' });
