+++ server/routes/auth.routes.js
@@ -1,42 +1,7 @@
 const express = require('express');
 const router = express.Router();
-const db = require('../db/db');
-const bcrypt = require('bcrypt');
-const jwt = require('jsonwebtoken');
+const authController = require('../controllers/authController');
 
-router.post('/login', (req, res) => {
-    const { emailOrPhone, password } = req.body;
-
-    if (!emailOrPhone || !password) {
-        return res.status(400).json({ error: 'Email or phone number and password are required' });
-    }
-
-    const sql = `
-      SELECT * FROM users 
-      WHERE email = ? OR phone_number = ?
-    `;
-
-    db.query(sql, [emailOrPhone, emailOrPhone], async (err, results) => {
-        if (err) return res.status(500).json({ error: err });
-
-        if (results.length === 0) {
-            return res.status(401).json({ error: 'User not found' });
-        }
-        const user = results[0];
-        const isMatch = await bcrypt.compare(password, user.password);
-
-        if (!isMatch) {
-            return res.status(401).json({ error: 'Incorrect password' });
-        }
-
-        const token = jwt.sign(
-            { user_id: user.user_id, role: user.role },
-            process.env.JWT_SECRET,
-            { expiresIn: '1h' }
-        );
-
-        res.json({ message: 'Login successful 🔐', token });
-    });
-});
+router.post('/login', authController.login);
 
 module.exports = router;
